+++ src/WPBT/WPBT_Help.php
@@ -27,6 +27,11 @@ class WPBT_Help {
 	 * @return void
 	 */
 	public function add_help_tabs() {
+		$current_screen = get_current_screen();
+		if ( false === strpos( $current_screen->id, 'wp-beta-tester' ) ) {
+			return false;
+		}
+
 		get_current_screen()->set_help_sidebar(
 			'<p><strong>' . __( 'For more information:' ) . '</strong></p>' .
 			'<p>' . __( '<a href="https://make.wordpress.org/core/handbook/testing/beta-testing/">Beta Testing</a>', 'wordpress-beta-tester' ) . '</p>'
@@ -51,8 +56,15 @@ class WPBT_Help {
 			array(
 				'id'      => 'beta/RC',
 				'title'   => __( 'Beta/RC', 'wordpress-beta-tester' ),
-				'content' => '<p>' . __( 'Blah, blah, blah ... yada, yada, yada', 'wordpress-beta-tester' ) . '</p>',
+				'content' => '<p>' . __( 'You must currently be running a <em>Point release nightly</em> or <em>Bleeding edge nightly</em> before the option to select the specific <em>Beta/RC</em> stream will display. Once selected you will only see an update notice when the next release, RC, or beta is available.', 'wordpress-beta-tester' ) . '</p>',
+			)
+		);
 
+		get_current_screen()->add_help_tab(
+			array(
+				'id'      => 'dashboard',
+				'title'   => __( 'Dashboard Widget', 'wordpress-beta-tester' ),
+				'content' => '<p>' . __( 'A dashboard widget is displayed when the plugin is active. It will contain links to milestone commits and filing a bug report. It may contain links to Dev Notes, the Field Guide, and beta/RC release posts.', 'wordpress-beta-tester' ) . '</p>',
 			)
 		);